begin implementing multi-game-mode code
Added interface IGameMode to define a common format for all game modes. May be expanded in future, but should allow a lot of the code in Round.cs (which is for implementing the Standard game mode) to be moved to a different file. This should greatly improve the ease with which new game modes can be added.