fix command interspersing / race condition
due to how CsNetLib2 is constructed, two threads calling simultaneous the SendCommand functions will have their commands melded together. This was discovered via race condition where the loop blocking the thread which eventually instantiates the next round finishes and proceeds to start the next round whilst code for informing clients of the winner of a game is executing, resulting both in the erroneous sending of RSTR/etc commands before or after a game is terminated. This is fixed by moving the assignment which causes the loop to end to the end of the round-winning/game-winning code.